Moved the right half of the screen that contains graphs to its own component for optimization and reorganization
Also optimized imports on some components
